chore(fonts): match cli structure part 2
