+++ packages/astro/src/assets/fonts/definitions.ts
@@ -1,6 +1,5 @@
 import type * as unifont from 'unifont';
 import type { CollectedFontForMetrics } from './core/optimize-fallbacks.js';
-/* eslint-disable @typescript-eslint/no-empty-object-type */
 import type {
 	AstroFontProvider,
 	FontFaceMetrics,
@@ -27,28 +26,6 @@ export interface RemoteFontProviderResolver {
 
 export interface LocalProviderUrlResolver {
 	resolve: (input: string) => string;
-}
-
-type SingleErrorInput<TType extends string, TData extends Record<string, any>> = {
-	type: TType;
-	data: TData;
-	cause: unknown;
-};
-
-export type ErrorHandlerInput =
-	| SingleErrorInput<
-			'cannot-load-font-provider',
-			{
-				entrypoint: string;
-			}
-	  >
-	| SingleErrorInput<'unknown-fs-error', {}>
-	| SingleErrorInput<'cannot-fetch-font-file', { url: string }>
-	| SingleErrorInput<'cannot-extract-font-type', { url: string }>
-	| SingleErrorInput<'cannot-extract-data', { family: string; url: string }>;
-
-export interface ErrorHandler {
-	handle: (input: ErrorHandlerInput) => Error;
 }
 
 interface ProxyData {
