+++ packages/astro/src/assets/fonts/infra/build-remote-font-provider-mod-resolver.ts
@@ -1,0 +1,9 @@
+import type { RemoteFontProviderModResolver } from '../definitions.js';
+
+export function createBuildRemoteFontProviderModResolver(): RemoteFontProviderModResolver {
+	return {
+		resolve(id) {
+			return import(id);
+		},
+	};
+}
