+++ packages/astro/src/assets/fonts/infra/build-url-proxy-hash-resolver.ts
@@ -1,0 +1,15 @@
+import type { Hasher, UrlProxyContentResolver, UrlProxyHashResolver } from '../definitions.js';
+
+export function createBuildUrlProxyHashResolver({
+	hasher,
+	contentResolver,
+}: {
+	hasher: Hasher;
+	contentResolver: UrlProxyContentResolver;
+}): UrlProxyHashResolver {
+	return {
+		resolve({ originalUrl, type }) {
+			return `${hasher.hashString(contentResolver.resolve(originalUrl))}.${type}`;
+		},
+	};
+}
