+++ packages/astro/src/assets/fonts/infra/capsize-font-metrics-resolver.ts
