+++ packages/astro/src/assets/fonts/infra/dev-remote-font-provider-mod-resolver.ts
@@ -1,14 +1,6 @@
 import type { ViteDevServer } from 'vite';
 import type { RemoteFontProviderModResolver } from '../definitions.js';
 
-export function createBuildRemoteFontProviderModResolver(): RemoteFontProviderModResolver {
-	return {
-		resolve(id) {
-			return import(id);
-		},
-	};
-}
-
 export function createDevServerRemoteFontProviderModResolver({
 	server,
 }: {
