+++ packages/astro/src/assets/fonts/infra/font-type-extractor.ts
@@ -1,21 +1,20 @@
 import { extname } from 'node:path';
-import type { ErrorHandler, FontTypeExtractor } from '../definitions.js';
+import { AstroError, AstroErrorData } from '../../../core/errors/index.js';
+import type { FontTypeExtractor } from '../definitions.js';
 import { isFontType } from '../utils.js';
 
-export function createFontTypeExtractor({
-	errorHandler,
-}: {
-	errorHandler: ErrorHandler;
-}): FontTypeExtractor {
+export function createFontTypeExtractor(): FontTypeExtractor {
 	return {
 		extract(url) {
 			const extension = extname(url).slice(1);
 			if (!isFontType(extension)) {
-				throw errorHandler.handle({
-					type: 'cannot-extract-font-type',
-					data: { url },
-					cause: `Unexpected extension, got "${extension}"`,
-				});
+				throw new AstroError(
+					{
+						...AstroErrorData.CannotExtractFontType,
+						message: AstroErrorData.CannotExtractFontType.message(url),
+					},
+					{ cause: `Unexpected extension, got "${extension}"` },
+				);
 			}
 			return extension;
 		},
