+++ packages/astro/src/assets/fonts/infra/xxhash-hasher.ts
@@ -2,7 +2,7 @@ import xxhash from 'xxhash-wasm';
 import type { Hasher } from '../definitions.js';
 import { sortObjectByKey } from '../utils.js';
 
-export async function createXxHasher(): Promise<Hasher> {
+export async function createXxhashHasher(): Promise<Hasher> {
 	const { h64ToString: hashString } = await xxhash();
 	return {
 		hashString,
