+++ packages/astro/test/units/assets/fonts/core.test.js
@@ -12,7 +12,6 @@ import {
 	createSpyUrlProxy,
 	fakeFontMetricsResolver,
 	fakeHasher,
-	simpleErrorHandler,
 } from './utils.js';
 
 describe('fonts core', () => {
@@ -347,7 +346,7 @@ describe('fonts core', () => {
 				normalizeRemoteFontFaces({
 					fonts: [],
 					urlProxy,
-					fontTypeExtractor: createFontTypeExtractor({ errorHandler: simpleErrorHandler }),
+					fontTypeExtractor: createFontTypeExtractor(),
 				}).length,
 				0,
 			);
@@ -380,7 +379,7 @@ describe('fonts core', () => {
 						},
 					],
 					urlProxy,
-					fontTypeExtractor: createFontTypeExtractor({ errorHandler: simpleErrorHandler }),
+					fontTypeExtractor: createFontTypeExtractor(),
 				}).length,
 				5,
 			);
@@ -405,7 +404,7 @@ describe('fonts core', () => {
 						src: [{ url: '/2', format: 'woff2' }],
 					},
 				],
-				fontTypeExtractor: createFontTypeExtractor({ errorHandler: simpleErrorHandler }),
+				fontTypeExtractor: createFontTypeExtractor(),
 			});
 			assert.deepStrictEqual(collected, [
 				{
@@ -456,7 +455,7 @@ describe('fonts core', () => {
 						],
 					},
 				],
-				fontTypeExtractor: createFontTypeExtractor({ errorHandler: simpleErrorHandler }),
+				fontTypeExtractor: createFontTypeExtractor(),
 			});
 			assert.deepStrictEqual(collected, [
 				{
@@ -506,7 +505,7 @@ describe('fonts core', () => {
 						src: [{ url: '/2', format: 'woff2' }, { name: 'Foo' }, { url: '/also-ignored.ttf' }],
 					},
 				],
-				fontTypeExtractor: createFontTypeExtractor({ errorHandler: simpleErrorHandler }),
+				fontTypeExtractor: createFontTypeExtractor(),
 			});
 			assert.deepStrictEqual(fonts, [
 				{
@@ -589,7 +588,7 @@ describe('fonts core', () => {
 						src: [{ url: '//example.com/font.woff2' }, { url: 'http://example.com/font.woff' }],
 					},
 				],
-				fontTypeExtractor: createFontTypeExtractor({ errorHandler: simpleErrorHandler }),
+				fontTypeExtractor: createFontTypeExtractor(),
 			});
 
 			assert.deepStrictEqual(fonts, [
